More refactor of package names
