+++ app/src/androidTest/java/io/lunarlogic/aircasting/MyAccountTest.kt
@@ -20,7 +20,7 @@ import io.lunarlogic.aircasting.helpers.getMockWebServerFrom
 import io.lunarlogic.aircasting.lib.Settings
 import io.lunarlogic.aircasting.networking.services.ApiServiceFactory
 import io.lunarlogic.aircasting.screens.new_session.LoginActivity
-import io.lunarlogic.aircasting.screens.settings.my_account.MyAccountActivity
+import io.lunarlogic.aircasting.screens.settings.clear_sd_card.my_account.MyAccountActivity
 import junit.framework.Assert.assertEquals
 import junit.framework.Assert.assertTrue
 import org.junit.After
