+++ app/src/main/java/io/lunarlogic/aircasting/screens/settings/clear_sd_card/ClearSDCardController.kt
@@ -20,8 +20,8 @@ import io.lunarlogic.aircasting.screens.new_session.connect_airbeam.TurnOnBlueto
 import io.lunarlogic.aircasting.screens.new_session.connect_airbeam.TurnOnLocationServicesViewMvc
 import io.lunarlogic.aircasting.screens.new_session.select_device.DeviceItem
 import io.lunarlogic.aircasting.screens.new_session.select_device.SelectDeviceViewMvc
-import io.lunarlogic.aircasting.screens.settings.SDCardCleared.SDCardClearedViewMvc
-import io.lunarlogic.aircasting.screens.settings.restart_airbeam.RestartAirBeamViewMvc
+import io.lunarlogic.aircasting.screens.settings.clear_sd_card.sd_card_cleared.SDCardClearedViewMvc
+import io.lunarlogic.aircasting.screens.settings.clear_sd_card.restart_airbeam.RestartAirBeamViewMvc
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.GlobalScope
 import kotlinx.coroutines.async
