+++ app/src/main/java/io/lunarlogic/aircasting/screens/settings/clear_sd_card/ClearSDCardWizardNavigator.kt
@@ -10,11 +10,11 @@ import io.lunarlogic.aircasting.screens.new_session.connect_airbeam.TurnOnLocati
 import io.lunarlogic.aircasting.screens.new_session.connect_airbeam.TurnOnLocationServicesViewMvc
 import io.lunarlogic.aircasting.screens.new_session.select_device.SelectDeviceFragment
 import io.lunarlogic.aircasting.screens.new_session.select_device.SelectDeviceViewMvc
-import io.lunarlogic.aircasting.screens.settings.SDCardCleared.SDCardClearedFragment
-import io.lunarlogic.aircasting.screens.settings.SDCardCleared.SDCardClearedViewMvc
-import io.lunarlogic.aircasting.screens.settings.clearing_sd_card.ClearingSDCardFragment
-import io.lunarlogic.aircasting.screens.settings.restart_airbeam.RestartAirBeamFragment
-import io.lunarlogic.aircasting.screens.settings.restart_airbeam.RestartAirBeamViewMvc
+import io.lunarlogic.aircasting.screens.settings.clear_sd_card.sd_card_cleared.SDCardClearedFragment
+import io.lunarlogic.aircasting.screens.settings.clear_sd_card.sd_card_cleared.SDCardClearedViewMvc
+import io.lunarlogic.aircasting.screens.settings.clear_sd_card.clearing_sd_card.ClearingSDCardFragment
+import io.lunarlogic.aircasting.screens.settings.clear_sd_card.restart_airbeam.RestartAirBeamFragment
+import io.lunarlogic.aircasting.screens.settings.clear_sd_card.restart_airbeam.RestartAirBeamViewMvc
 
 class ClearSDCardWizardNavigator(
     private val mViewMvc: ClearSDCardViewMvc,
