+++ app/src/main/java/io/lunarlogic/aircasting/screens/settings/clear_sd_card/clearing_sd_card/ClearingSDCardViewMvc.kt
@@ -1,0 +1,4 @@
+package io.lunarlogic.aircasting.screens.settings.clear_sd_card.clearing_sd_card
+
+interface ClearingSDCardViewMvc{
+}
