+++ app/src/main/java/io/lunarlogic/aircasting/screens/settings/clear_sd_card/my_account/MyAccountActivity_MembersInjector.java
@@ -1,0 +1,32 @@
+// Generated by Dagger (https://dagger.dev).
+package io.lunarlogic.aircasting.screens.settings.clear_sd_card.my_account;
+
+import dagger.MembersInjector;
+import dagger.internal.InjectedFieldSignature;
+import io.lunarlogic.aircasting.lib.Settings;
+import javax.inject.Provider;
+
+@SuppressWarnings({
+    "unchecked",
+    "rawtypes"
+})
+public final class MyAccountActivity_MembersInjector implements MembersInjector<MyAccountActivity> {
+  private final Provider<Settings> settingsProvider;
+
+  public MyAccountActivity_MembersInjector(Provider<Settings> settingsProvider) {
+    this.settingsProvider = settingsProvider;
+  }
+
+  public static MembersInjector<MyAccountActivity> create(Provider<Settings> settingsProvider) {
+    return new MyAccountActivity_MembersInjector(settingsProvider);}
+
+  @Override
+  public void injectMembers(MyAccountActivity instance) {
+    injectSettings(instance, settingsProvider.get());
+  }
+
+  @InjectedFieldSignature("io.lunarlogic.aircasting.screens.settings.clear_sd_card.my_account.MyAccountActivity.settings")
+  public static void injectSettings(MyAccountActivity instance, Settings settings) {
+    instance.settings = settings;
+  }
+}
