+++ app/src/main/java/io/lunarlogic/aircasting/screens/settings/clear_sd_card/my_account/MyAccountController.kt
@@ -1,4 +1,4 @@
-package io.lunarlogic.aircasting.screens.settings.my_account
+package io.lunarlogic.aircasting.screens.settings.clear_sd_card.my_account
 
 import android.content.Context
 import io.lunarlogic.aircasting.database.DatabaseProvider
