+++ app/src/main/java/io/lunarlogic/aircasting/screens/settings/clear_sd_card/restart_airbeam/RestartAirBeamViewMvcImpl.kt
@@ -1,4 +1,4 @@
-package io.lunarlogic.aircasting.screens.settings.restart_airbeam
+package io.lunarlogic.aircasting.screens.settings.clear_sd_card.restart_airbeam
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
