+++ app/src/main/java/io/lunarlogic/aircasting/screens/settings/clear_sd_card/sd_card_cleared/SDCardClearedFragment.kt
@@ -1,4 +1,4 @@
-package io.lunarlogic.aircasting.screens.settings.SDCardCleared
+package io.lunarlogic.aircasting.screens.settings.clear_sd_card.sd_card_cleared
 
 import android.os.Bundle
 import android.view.LayoutInflater
