--- app/src/main/java/io/lunarlogic/aircasting/screens/settings/clearing_sd_card/ClearingSDCardViewMvc.kt
@@ -1,4 +1,0 @@
-package io.lunarlogic.aircasting.screens.settings.clearing_sd_card
-
-interface ClearingSDCardViewMvc{
-}
